tidying

+++ scripts/install_webserver.sh
@@ -1,22 +1,12 @@
 #!/usr/bin/env bash
 source /usr/local/bootstrap/var.env
 
-IFACE=`route -n | awk '$1 == "192.168.2.0" {print $8}'`
-CIDR=`ip addr show ${IFACE} | awk '$2 ~ "192.168.2" {print $2}'`
-IP=${CIDR%%/24}
-
-if [ -d /vagrant ]; then
-  LOG="/vagrant/consul_${HOSTNAME}.log"
-else
-  LOG="consul.log"
+# Idempotency hack - if this file exists don't run the rest of the script
+if [ -f "/var/vagrant_web_server" ]; then
+   exit 0
 fi
 
-# Idempotency hack - if this file exists don't run the rest of the script
-#if [ -f "/var/vagrant_web_server" ]; then
-#    exit 0
-#fi
-
-#touch /var/vagrant_web_server
+touch /var/vagrant_web_server
 
 # remove nginx default website
 sudo rm -f /etc/nginx/sites-enabled/default
